Moved profile and base config classes to their own file
