+++ src/zenml/cli/utils.py
@@ -21,7 +21,7 @@ from dateutil import tz
 from rich import box, table
 from rich.text import Text
 
-from zenml.config.global_config import ProfileConfiguration
+from zenml.config.profile_config import ProfileConfiguration
 from zenml.console import console
 from zenml.constants import IS_DEBUG_ENV
 from zenml.enums import StackComponentType
